production_requirements.txt
annotated-types
anyio
certifi
click
distro
exceptiongroup
fastapi
h11
httpcore
httptools
httpx
idna
openai
pydantic
pydantic_core
python-dotenv
PyYAML
sniffio
starlette
tqdm
typing_extensions
uvicorn
uvloop
watchfiles
websockets
